Block, Turnouts and Sensors Scripts Refined
The scripts for blocks, turnouts and sensors have been separated out. Based on the current setup the scripts for blocks and turnouts will need to be run first to create the objects. Then the sensor script will be run which will iterate through all the objects and create the necessary sensors. At this time no check is being done to make sure that existing objects are not being overwritten.