MML Panel Lights Phase 1
This phase deals with the script to generate the panel lights for MML. This also deals with the kind of information to send the panel board for blocks with switches. In this version every block that is configured to show a panel light is given a light on the panel. Those with switches are given an additional light that indicates the state of the turnout. The idea behind this is to send the Arduino code messages about the state of the entire block using the bits for the lightsthere by giving the Arduino code the following states of information:

- Switch Normal, Block Occupied
- Switch Normal, Block Unoccupied
- Switch Thrown, Block Occupied
- Switch Thrown, Block Unoccupied

The Arduino code will be responsible for configuring the color and the lit status of the lights for this setup based on the two bit of information received.