MML Signals Script 90% Complete
The Arduino sketch for MML is nearly finalized. The script needs documentation regarding the pin to bit mapping.

Signal Aspects for RY two aspect signaling had to be updated because of the way JMRI double output signal heads work. Instead of using "yellow" in the appearance file, we are using "green" so that the right light is lit. When constructing the signal a yellow led will be used though.